test(workmanager): add singleton test and dummy mocked workmanager test
